requirements.txt
ccxt
python-dotenv
requests
pandas
numpy
scikit-learn
joblib
python-telegram-bot
matplotlib
seaborn